Refactor actions, reducers and action consts into their own files
