+++ client/react/app/actions/actionConstants.ts
@@ -1,0 +1,36 @@
+export const LIST_VIEW = 'LIST_VIEW';
+export const CHART_VIEW = 'CHART_VIEW';
+
+export const UPDATE_VIEW = 'UPDATE_VIEW';
+export const SELECT_DATE = 'SELECT_DATE';
+
+export const REQUEST_SCORECARDS = 'REQUEST_SCORECARDS';
+export const RECEIVE_SCORECARDS = 'RECEIVE_SCORECARDS';
+export const REQUEST_SCORECARD = 'REQUEST_SCORECARD';
+export const RECEIVE_SCORECARD = 'RECEIVE_SCORECARD';
+export const UPDATE_SCORECARD = 'RECEIVE_SCORECARD';
+
+export const REQUEST_CHECKLISTS = 'REQUEST_CHECKLISTS';
+export const RECEIVE_CHECKLISTS = 'RECEIVE_CHECKLISTS';
+export const REQUEST_CHECKLIST = 'REQUEST_CHECKLIST';
+export const RECEIVE_CHECKLIST = 'RECEIVE_CHECKLIST';
+export const UPDATE_CHECKLIST = 'UPDATE_CHECKLIST';
+
+export const REQUEST_ENTRIES = 'REQUEST_ENTRIES';
+export const RECEIVE_ENTRIES = 'RECEIVE_ENTRIES';
+export const REQUEST_ENTRY = 'REQUEST_ENTRY';
+export const RECEIVE_ENTRY = 'RECEIVE_ENTRY';
+
+export const REQUEST_CURRENT_USER = 'REQUEST_CURRENT_USER';
+export const RECEIVE_CURRENT_USER = 'RECEIVE_CURRENT_USER';
+
+export const REQUEST_CHALLENGES = 'REQUEST_CHALLENGES';
+export const RECEIVE_CHALLENGES = 'RECEIVE_CHALLENGES';
+export const REQUEST_ALL_CHALLENGES = 'REQUEST_ALL_CHALLENGES';
+export const RECEIVE_ALL_CHALLENGES = 'RECEIVE_ALL_CHALLENGES';
+export const REQUEST_MY_CHALLENGES = 'REQUEST_MY_CHALLENGES';
+export const RECEIVE_MY_CHALLENGES = 'RECEIVE_MY_CHALLENGES';
+export const UPDATE_CHALLENGE = 'UPDATE_CHALLENGE';
+
+export const REQUEST_ALL_STATS = 'REQUEST_ALL_STATS';
+export const RECEIVE_ALL_STATS = 'RECEIVE_ALL_STATS';
