+++ client/react/app/components/dashboard/Dashboard.less
@@ -49,7 +49,7 @@
   min-height: 400px;
   padding-left: 40px;
   padding-top: 8px;
-  width: 90%;
+  width: 100%;
 }
 
 .dashboard-preview-container {
