+++ client/react/app/components/dashboard/Today.tsx
@@ -187,18 +187,24 @@ class Today extends React.Component<TodayProps, TodayState> {
 
         <div className='default-container'>
           <div className='clearfix'>
-            <div className='dashboard-preview-container pull-left'>
+            {/* <div className='dashboard-preview-container pull-left'>
+            <DashboardPreview
+                selected={selected}
+                handleScorecardClicked={this.createNewScorecard.bind(this)}
+                handleChecklistClicked={this.createNewChecklist.bind(this)}
+                handleEntryClicked={this.createNewEntry.bind(this)} />
+
             <TodayProgressBar
                 currentScore={currentScore}
                 averageScore={averageScore}
                 topScore={topScore} />
-              {/* <DashboardPreview
-                selected={selected}
-                handleScorecardClicked={this.createNewScorecard.bind(this)}
-                handleChecklistClicked={this.createNewChecklist.bind(this)}
-                handleEntryClicked={this.createNewEntry.bind(this)} /> */}
-            </div>
-            <div className='dashboard-scorecard-container pull-right'>
+            </div> */}
+            <div className='dashboard-scorecard-container block-center'>
+            <TodayProgressBar
+                currentScore={currentScore}
+                averageScore={averageScore}
+                topScore={topScore} />
+
               <Scorecard
                 tasks={tasks}
                 challenges={challenges}
